+++ src/main/java/com/example/bobfriend/model/dto/comment/Create.java
@@ -1,0 +1,17 @@
+package com.example.bobfriend.model.dto.comment;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.validation.constraints.NotBlank;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Create {
+    @NotBlank
+    private String content;
+}
